Moves single_arg utility to lib.args
